move helper functions to common file
